+++ demo/src/app/components/+alert/strong/strong.component.ts
@@ -1,0 +1,9 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+    selector: 'demo-alert-strong',
+    templateUrl: './strong.component.html'
+})
+export class DemoAlertStrongComponent implements OnInit {
+    ngOnInit() {}
+}