+++ demo/src/app/components/+alert/weak/weak.component.ts
@@ -1,0 +1,9 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+    selector: 'demo-alert-weak',
+    templateUrl: './weak.component.html'
+})
+export class DemoAlertWeakComponent implements OnInit {
+    ngOnInit() {}
+}