+++ demo/src/app/components/+input/demo/size/size.component.ts
@@ -1,0 +1,17 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+    selector: 'demo-input-size',
+    templateUrl: 'size.component.html'
+})
+export class DemoInputSizeComponent implements OnInit {
+    public value;
+
+    constructor() {}
+
+    ngOnInit() {}
+
+    public enter() {
+        console.log('enter', this.value);
+    }
+}