+++ src/module.ts
@@ -52,6 +52,7 @@ import { ThyMarkdownModule } from './markdown';
 import { ThyFlexibleTextModule } from './flexible-text';
 import { ThyDragDropModule } from './drag-drop/module';
 import { ThySelectCommonModule } from './core/select/module';
+import { ThySkeletonModule } from './skeleton';
 
 const IMPORT_EXPORT_MODULES = [
     BrowserAnimationsModule,
@@ -107,7 +108,8 @@ const IMPORT_EXPORT_MODULES = [
     ThyMarkdownModule,
     ThyFlexibleTextModule,
     ThyDragDropModule,
-    ThySelectCommonModule
+    ThySelectCommonModule,
+    ThySkeletonModule
 ];
 
 @NgModule({