+++ CHANGELOG.md
@@ -1,3 +1,76 @@
+## [7.6.1](https://github.com/worktile/ngx-tethys/compare/7.6.0...7.6.1) (2019-11-13)
+
+
+### Bug Fixes
+
+* **custom-select:** fix class name for options ([3ca5d26](https://github.com/worktile/ngx-tethys/commit/3ca5d26))
+* **custom-select:** remove checked-icon if ([f3b1040](https://github.com/worktile/ngx-tethys/commit/f3b1040))
+
+
+### Features
+
+* **alert:** add strong type of danger & secondary ([2b4ba88](https://github.com/worktile/ngx-tethys/commit/2b4ba88))
+
+
+
+# [7.6.0](https://github.com/worktile/ngx-tethys/compare/7.5.29...7.6.0) (2019-11-12)
+
+
+### Bug Fixes
+
+* **custom-select:** fix test case and stop close when muitiple is true ([2fe45a7](https://github.com/worktile/ngx-tethys/commit/2fe45a7))
+* **select-control:** fix change newlines when selected item is overflow ([65c33a3](https://github.com/worktile/ngx-tethys/commit/65c33a3))
+
+
+### Features
+
+* **custom-select:** support clear when panel is open ([4944088](https://github.com/worktile/ngx-tethys/commit/4944088))
+* add cdkScrollable for demo ([b1a8f9c](https://github.com/worktile/ngx-tethys/commit/b1a8f9c))
+* **skeleton:** add skeleton module contains thy-skeleton comp #INF-124 ([a49d883](https://github.com/worktile/ngx-tethys/commit/a49d883)), closes [#INF-124](https://github.com/worktile/ngx-tethys/issues/INF-124)
+
+
+
+## [7.5.29](https://github.com/worktile/ngx-tethys/compare/7.5.27...7.5.29) (2019-11-07)
+
+
+### Bug Fixes
+
+* **action-menu:** change spacing between icon and text ([0a8e68c](https://github.com/worktile/ngx-tethys/commit/0a8e68c))
+* **custom-select:** fix option active and hover conflict ([ed4d8b5](https://github.com/worktile/ngx-tethys/commit/ed4d8b5))
+* **list:** update list-item bg to f3f3f3 ([9b1729f](https://github.com/worktile/ngx-tethys/commit/9b1729f))
+* **select:** fix option paddingRight from 35 to 20 ([fa7d1cd](https://github.com/worktile/ngx-tethys/commit/fa7d1cd)), closes [#INF-94](https://github.com/worktile/ngx-tethys/issues/INF-94)
+* **tree:** set default tree type ([8d40d02](https://github.com/worktile/ngx-tethys/commit/8d40d02))
+* **tree-select:** modify style of single selection ([5694019](https://github.com/worktile/ngx-tethys/commit/5694019))
+
+
+### Features
+
+* **input:** input default autocomplete off (#INF-133,#AE-140) ([3cdb388](https://github.com/worktile/ngx-tethys/commit/3cdb388)), closes [#INF-133](https://github.com/worktile/ngx-tethys/issues/INF-133) [#AE-140](https://github.com/worktile/ngx-tethys/issues/AE-140)
+
+
+
+## [7.5.28](https://github.com/worktile/ngx-tethys/compare/7.5.27...7.5.28) (2019-10-25)
+
+
+### Bug Fixes
+
+* **action-menu:** change spacing between icon and text ([0a8e68c](https://github.com/worktile/ngx-tethys/commit/0a8e68c))
+* **list:** update list-item bg to f3f3f3 ([9b1729f](https://github.com/worktile/ngx-tethys/commit/9b1729f))
+* **select:** fix option paddingRight from 35 to 20 ([fa7d1cd](https://github.com/worktile/ngx-tethys/commit/fa7d1cd)), closes [#INF-94](https://github.com/worktile/ngx-tethys/issues/INF-94)
+
+
+
+## [7.5.27](https://github.com/worktile/ngx-tethys/compare/7.5.25...7.5.27) (2019-10-17)
+
+
+### Bug Fixes
+
+* remove left style for selected-value ([c922c6d](https://github.com/worktile/ngx-tethys/commit/c922c6d))
+* **input:** add input size variables ([9470a2e](https://github.com/worktile/ngx-tethys/commit/9470a2e))
+* delete nested span when show custom option ([d196d2a](https://github.com/worktile/ngx-tethys/commit/d196d2a))
+
+
+
 ## [7.5.26](https://github.com/worktile/ngx-tethys/compare/7.5.25...7.5.26) (2019-10-14)